requirements.txt
# DipMaster Trading System Dependencies
# Production requirements for the DipMaster AI Trading System

# ============================================================================
# CORE TRADING DEPENDENCIES
# ============================================================================

# Data Processing & Analysis
pandas>=2.0.0,<3.0.0          # DataFrame operations and data manipulation
numpy>=1.24.0,<2.0.0          # Numerical computations and array operations

# Cryptocurrency Exchange Integration
python-binance>=1.0.19,<2.0.0 # Binance API client for trading operations
websocket-client>=1.6.0,<2.0.0 # WebSocket client for real-time data streams
aiohttp>=3.9.0,<4.0.0          # Async HTTP client for API calls

# Technical Analysis
ta>=0.10.2,<1.0.0              # Technical analysis indicators (RSI, MA, etc.)

# ============================================================================
# SYSTEM & INFRASTRUCTURE
# ============================================================================

# Configuration & Environment
python-dotenv>=1.0.0,<2.0.0   # Environment variable management
pydantic>=2.4.0,<3.0.0         # Data validation and settings management

# Logging & Monitoring  
rich>=13.5.0,<14.0.0           # Rich console output and logging
prometheus-client>=0.18.0,<1.0.0 # Metrics collection and monitoring

# ============================================================================
# WEB INTERFACE & API
# ============================================================================

# Web Framework (for dashboard)
fastapi>=0.104.0,<1.0.0        # Fast API framework for REST endpoints
uvicorn>=0.24.0,<1.0.0         # ASGI server for FastAPI applications

# ============================================================================
# DATA VISUALIZATION (Optional)
# ============================================================================

# Charting & Analysis (optional, for analysis mode)
matplotlib>=3.7.0,<4.0.0       # Plotting and chart generation
seaborn>=0.12.0,<1.0.0         # Statistical data visualization

# ============================================================================
# PERSISTENCE & CACHING (Optional)
# ============================================================================

# Database (optional, for advanced setups)
sqlalchemy>=2.0.0,<3.0.0       # SQL toolkit and ORM

# Caching (optional, for performance optimization)
redis>=5.0.0,<6.0.0            # In-memory caching for hot data

# ============================================================================
# SCHEDULING & UTILITIES
# ============================================================================

# Task Scheduling
schedule>=1.2.0,<2.0.0         # Simple job scheduling